use correct dir

+++ test/cleanup-branches.sh
@@ -3,6 +3,12 @@
 REPO="$1"
 
 cd "${REPO}" || exit 1
-set -e
-git push --delete origin env-dev
-git push --delete origin deploy-pr/env-dev
+set +e
+if git ls-remote --exit-code --heads origin refs/heads/env-dev; then
+    set -e
+    git push --delete origin env-dev
+fi
+if git ls-remote --exit-code --heads origin refs/heads/deploy-pr/env-dev; then
+    set -e
+    git push --delete origin deploy-pr/env-dev
+fi
